share code block logic
